cpu/percpu: introduce interior mutability in PerCpuUnsafe
The PerCpuUnsafe struct is reentrant, meaning multiple references can
be acquired to it from several callsites in a single call stack.
Since mutable references must always be unique, use only shared
references and add interior mutability to the structure via Cell and
RefCell. We use Cell for trivial Copy types, and RefCell for more
complex types that we need to give references to.

This allows us to remove dynamic borrow checking on the PerCpu struct
via RefCell as a whole, as we never give out mutable references,
making every reference automatically safe.

Signed-off-by: Carlos López <carlos.lopez@suse.com>