cpu/percpu: remove PerCpuUnsafe
Now that the PerCpu and PerCpuUnsafe structures have interior
mutability, it's always safe to acquire a reference to them, since
they are always CPU-local and never modified outside the current CPU.
Thus, we can remove PerCpuUnsafe and inline everything into a single
PerCpu structure. This also removes a lot of unnecessary unsafe code,
simplifying the codebase.

Signed-off-by: Carlos López <carlos.lopez@suse.com>