cpu/percpu: hold safe reference to the GHCB page
Now that the PerCpu is has interior mutability, and that the GHCB does
as well, it is safe to hold references to them, as they can never
mutably aliased. Thus, there is no need to store a const pointer and
the unsafe code around it.

This obsoletes the GHCBRef type, as there is no longer a need for a
wrapper.

Signed-off-by: Carlos López <carlos.lopez@suse.com>