cpu/percpu: remove RWLock around runqueue
The PerCpu structure can only be accessed from the current CPU, so
there is no need for locking. Moreover, if the runqueue is acquired
with write access more than once this will cause a deadlock. Replace
the RWLock with a RefCell. This way, if we attempt to acquire the
runqueue with write access more than once we will panic immediately.

Signed-off-by: Carlos López <carlos.lopez@suse.com>