igvmbuilder: Capture the IGVM firmware memory map parameter
When loading firmware from an IGVM file, capture the address range of
the IGVM memory map parameter into the firmware info block instead of
emitting it as an IGVM parameter.  This prevents the host loader from
supplying the memory map directly to the firmware, and instead enables
the SVSM to supply a trusted memory map directly into firmware memory.

Signed-off-by: Jon Lange <jlange@microsoft.com>