igmvbuilder/ovmf: allow alternative fw address
On a TD Partitioning platform, guest boot firmware is not located to end
at 4GB. This is to avert a page collision at the top of 4GB, where the
stage1 trampoline page is placed for bootstrapping SVSM itself.

Introduce OVMF_RESET_VECTOR_GUID, a new OVMF GUIDed structure describing
the location of the firmware's expected reset vector as well as the
compatibility mask it supports. If this structure is present in the OVMF
image, alter the firmware's starting address based on this reset vector
location and the firmware's compatibility mask. By default, OVMF does
not support TDP and therefore OVMF_RESET_VECTOR_GUID must be used to
indicate a TDP-compatible OVMF image. The default starting address
remains unchanged.

Signed-off-by: Peter Fang <peter.fang@intel.com>