kernel/mm/pagetable: Add PTWalkAttr to check pte access rights
Whether an access is permitted by a translation is determined by the
access rights specified by the paging-structure entries controlling the
translation: paging-mode modifiers in CR0, CR4, and the IA32_EFER MSR;
EFLAGS.AC; the supervisor/user mode access.

Add PTWalkAttr to include all these factors and implement function
check_access_rights() to validate the access rights of a pte. A Ok
containing the translated physical address and a bool flag representing
if the translated PTEntry is leaf or not will be returned if it passes
the validation, otherwise returns Err with the corresponding page fault
error code flags.

The access rights validation will be used when traversals a page-table
to translate a virtual address.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>