kernel/mm/ptguard: Introduce MemMappingGuard for temporary memory mappings
MemMappingGuard is introduced to represent a temporary memory mapping
for a specified memory range. It provides read and write functions that
facilitate data access within the mapped memory using specified offsets
and lengths. When a MemMappingGuard instance is dropped, it will
automatically unmaps the associated memory range.

Additionally, the InsnMachineMem trait has been implemented for
MemMappingGuard, enabling it to support memory access operations
required by the instruction decoder. This is essential for the emulation
of specific instructions.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>