kernel/mm/ptguard: Add MemMappingGuard for temporary memory mapping
Introduce MemMappingGuard to represent a temporary memory mapping. It
contains a PerCPUPageMappingGuard, which is extended to remap multiple
non-contigous 4k pages to a contigous virtual address space. The
MemMappingGuard provides read and write that facilitate a specific data
type access within the remapped memory using a specified offset. When a
MemMappingGuard instance is dropped, the PerCPUPageMappingGuard is also
dropped which will unmap the associated memory range.

Additionally, the InsnMachineMem trait has been implemented for
MemMappingGuard, enabling it to support memory access operations
required by the instruction decoder. This is essential for the emulation
of specific instructions.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>