kernel/mm/pagetable: Add virtual addr translation supprot
Introduce translate_addr() function for PageTable to translate a virtual
address. During the translation, the PTEntry in each level is validated
against the memory access mode (write/fetch) and PTWalkAttr make sure
the pte is valid. An Ok is returned which contains the corresponding
physical address and the size of the page frame if the validation is
passed. Otherwise returns an Err which contains the corresponding page
fault error code.

Another fuction translate_addr_region() is similar, which translates a
virtual address region instead of a single virtual address. The access
rights are also validated during the translation. A successful
translation will return a vector contains the translated physical
addresses of 4k pages, otherwise returns Err which contains the
corresponding virtual address together with the page fault error code.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>