kernel/insn_decode: Enhance test cases for IOIO instructions
Enhance the TestCtx to implement the required methods of InsnMachineCtx.
Also introduce TestMem structure which implements InsnMachineMem trait.
With these, extend the test to support testing decoding of ins and outs
instructions, as well as emulation for ins/outs/in/out instructions.
These new tests ensure support for byte, word and doubleword IO port
size.

The instruction fuzz is also enhanced to test the emulate interface.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>