kernel/insn_decode: Emulate in/out instructions
DecodeInsnCtx has been enhanced to support the emulation of in and out
instructions. The in instruction is emulated by fetching data from the
IO port and writing it to the EAX/AX/AL register, while the out
instruction is emulated by reading data from the EAX/AX/AL register and
sending it to the IO port. Prior to emulation, the IO permission bitmap
is consulted to ensure proper access rights.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>