cpu/percpu: store LocalApic in an Option
Instead of having a separate boolean variable indicating whether APIC
emulation is enabled, simply store the LocalApic type in an Option.
When APIC emulation is enabled, the field in the PerCpu structure
will contain a Some(LocalApic), and a None otherwise. This enforces
at the type level a consistent state of APIC emulation and removes
duplicate checks for such state.

Signed-off-by: Carlos López <carlos.lopez@suse.com>