cpu/tss: fix TSS layout
According to the AMD manual, there is a reserved 64 bit entry between
the RSP and IST stack entries. Thus, separate it as such in the
X86Tss struct. While we are at it, mark it as packed(4), which is the
correct alignment, and add a layout test.

Even if there is a reserved entry, the first actual IST entry is
indexed by 1 and not 0, so use a NonZero type for IST indexing.

Signed-off-by: Carlos López <carlos.lopez@suse.com>