scripts: add serial port to communicate with tests running in SVSM
Add a serial port (COM3) to QEMU when we run unit tests in SVSM.
This way we can use it to communicate with tests running in the guest.

The serial port is connected to 2 pipes used for IN and OUT.
Pipes are created in a temporary folder and destroyed when
testing is completed.

The guest sends a request (1 byte) and the host responds according
to the type of the request. See the next commit for an example.

Signed-off-by: Stefano Garzarella <sgarzare@redhat.com>