kernel/testing: add svsm_test_io() method
The first test calling the svsm_test_io() function will initialize the
serial port (COM4) and use it exclusively, since it will be protected
by LockGuard.

This serial port can be used in tests running in SVSM to communicate
with the host. The protocol is simple, guest writes 1 byte identifying
the request (IORequest), subsequent in/out depends on each request.

Next commit will extend scripts/test-in-svsm.sh to handle the requests
in the host side.

Signed-off-by: Stefano Garzarella <sgarzare@redhat.com>