greq/services: add test to check the pre-calculated launch measurement
Ask the host for the pre-calculated launch measurement, and ask for a
VMPL0 attestation report and compare them.

This way we can avoid regressions in our tools for calculating the
expected launch measurement.

`scripts/test-in-svsm.sh` is now using `bin/igvmmeasure`, so let's
add `$(IGVMMEASUREBIN)` dependency to `test-in-svsm` target in the
Makefile.

Signed-off-by: Stefano Garzarella <sgarzare@redhat.com>