stage1: support TDX mode at reset vector
TD CPUs always start execution in protected (32-bit) non-paged mode at
the reset vector. Add the code necessary for a vCPU to run in the
resetvector section directly in 32-bit mode. Since there is no need for
a TD CPU to transition to 32-bit mode, it takes a slightly simpler path
to reach startup_32. TD CPUs are initialized with a flat 32-bit code
segment so absolute addressing can be used.

Signed-off-by: Peter Fang <peter.fang@intel.com>