stage1: support TDX in trampoline-only mode
Since TDP is the only platform right now that uses stage1-trampoline,
only allow platform type TDP when in trampoline-only mode. All TD vCPUs
start execution in 32-bit mode concurrently and therefore only TD BSP is
allowed to use the stage2 stack. Set vtom in Stage2LaunchInfo based on
the initial value of %ebx (GPAW).

Program an invalid address (0x7ffff000) into a TD AP's ESP register.
This address is chosen because a) it is most likely not an MMIO address
and b) it is an unaccepted page if it's backed by memory.

Signed-off-by: Peter Fang <peter.fang@intel.com>