error: subsume ApicError into SvsmError
This change removes `ApicError` as a type associated with APIC emulation
and defines it as a subordinate type of `SvsmError`, with variants
describing the types of errors that may be associated with APIC handling
(emulation or otherwise).  This permits unification of error types
across different components so they can all use `SvsmError`.

Signed-off-by: Jon Lange <jlange@microsoft.com>