dm: Add initial device model project
The device model is a user mode program which manages VM's life cycles
and handles various vmexits events. The initial version just calls exit()
syscall without doing anything else.

As this is a user mode program, not sure what is the preferred way to
put this project, just put it as a sub folder in coconut-svsm and in the
workspace. If in the workspace is not preferred, this can be changed in
the future.

Co-developed-by: Vijay Dhanraj <vijay.dhanraj@intel.com>
Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>