mm/pagetable: use PageBox for memory management
Remove calls to the low-level page allocation APIs
(allocate_zeroed_page, free_page) and replace them with uses of the
new PageBox type, which simplifies code in several places and reduces
the amount of unsafe code.

In order to improve readability, move the management of pages in the
pagetable into PTPage methods.

Signed-off-by: Carlos López <carlos.lopez@suse.com>