kernel/cpu/vc: Remove test_rdmsr_debug_ctl as not intercepted anymore
The commit b7e4be0a224fe5 in the kernel introduces a change that means
that sev-es and sev-snp guests do not now intercept the debug ctl MSR.
This means for kernels that include this commit this test will always
fail.

This commit removes the test. The read_msr functionality is tested by
the remaining test_rdmsr_apic() test case.

Signed-off-by: Roy Hopkins <roy.hopkins@suse.com>