igvmbuilder: fix incorrect CPUID table construction
The IGVM file specifies the set of CPUID leaves that are expected to be
provided when the IGVM file is loaded.  The builder was erroneously
specifying leaf EAX=1 with subleaf ECX=1, which is not architecturally
defined.  This caused the subsequent CPUID lookup to fail when the SVSM
would execute its CPUID lookup code.

Signed-off-by: Jon Lange <jlange@microsoft.com>