build: Allow building without mstpm
The svsm crate allows the libmstpm dependency to be optional, but since
it's a default feature there's no actual way to use make to compile it
that disables the feature. Add the necessary parts so that it can be
compiled out. This allows running FEATURES='""' make to not build the
mstpm.

Signed-off-by: Adam Dunlap <acdunlap@google.com>