platform: add GPA to validate_page_range()
validate_page_range() currently takes GVA due to the design of
PVALIDATE. In the case of TDX, GPA is required for page acceptance.

Add an additional input, paddr (PhysAddr), to validate_page_range(). An
explicit input for GPA is required because virt_to_phys() is not always
available. The input region may be out of range for the current
KERNEL_MAPPING.

Signed-off-by: Peter Fang <peter.fang@intel.com>