mm/validate: do not use NULL pointers in valid bitmap
At the moment, we initialize a global variable with a structure
(`ValidBitmap`) that holds a raw pointer to a bitmap. To signal that
the bitmap has not been initialized, we store a NULL pointer and
check on every access. Instead of doing this, simply store an
`Option<ValidBitmap>` in the global variable. If the bitmap has not
been initialized, the `Option` will be `None`, which we are forced to
check, avoiding potential NULL pointer dereferences and simplyfing
some of the methods of `ValidBitmap`.

Signed-off-by: Carlos López <carlos.lopez@suse.com>