stage2: Obtain CPUID and secrets page address from stage2 parameters
The current code captures the CPUID page and secrets page addresses in
multiple places, including IGVM parameters and hard-coded addresses.  It
is preferable to capture the addresses in exactly one place, the stage2
launch parameters (which are constructed either by stage1 or in the IGVM
file), so other code that requires those addresses can obtain them from
a single place.

Signed-off-by: Jon Lange <jlange@microsoft.com>