stage2: Relocate stage1/stage2 above 8 MB
The stage2 binary has grown beyond the point that it can fit below 640
KB with its heap.  The entirety of stage1, stage2, and all of the
initial kernel images, are relocated to a base address of 8 MB.  Stage 1
and the initial stage 2 stack consume the first 24 KB, followed by the
secrets page and CPUID page, and then come the stage 2 image (with
enough room for a heap) and the kernel images.

Signed-off-by: Jon Lange <jlange@microsoft.com>