kernel/insn_decode: Support for ins instruction
Add instruction decoding logic to support for the ins instruction. The
emulation of ins is achieved by reading data from an IO port and then
writing it to the decoded memory location.

Extend InsnMachineCtx to include a new method to manage IO in operations.
Additionally, the InsnMachineMem trait has been extended with a method
write_integer for writing integer data to memory, which is crucial for
the accurate emulation of the ins instruction.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>