stage2: Use low memory for the stage2 heap
Separating the stage2 heap into low memory and away from the stage2
binary means that the stage2 binary can grow to be arbitrarily large
without affecting the amount of memory available for heap usage.  The
area below 640 KB is never populated with initial data for the guest OS,
so it is free for stage 2 to use, and it is returned to a non-validated
state by the time the guest OS begins running.

Signed-off-by: Jon Lange <jlange@microsoft.com>