kernel/task/schedule: Disable interrupt during context switching
The instructions "mov %rdx, %cr3" and "movq (%rdi), %rsp" will switch
the page table and stack. An interrupt happening in the middle of the
switching can corrupt the stack, as it will push data to the old rsp
address but is translated by the new page table. To avoid such stack
corruption, disable interrupt before the switching the page table and
stack. The interrupt will be restored by the popfq.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>