kernel/cpu: Introduce IRQ state handling infrastructure
Introduce low-level helpers for IRQ managment and the IrqState struct.
This struct will be used to track per-cpu IRQ states and counts the
nesting level of irq_disable() calls. It will keep track of the IRQ
state when the nesting level was 0 and restores is when the level
reaches 0 again.

Co-authored-by: Tom Dohrmann <Erbse.13@gmx.de>
Signed-off-by: Joerg Roedel <jroedel@suse.de>