doc/design: Add design doc for object
The syscalls design philosophy is to provide a unified interface for the
user to access system resources, which are represented by object
handles. This makes objects a fundamental concept in the COCONUT-SVSM
kernel.

This document describes the object and object handle from both the user
mode's and the COCONUT-SVSM kernel's point of view.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>