mm: improve soundness of PageRef methods
Previously PageRef had AsMut<[u8; PAGE_SIZE]> and AsRef<[u8; PAGE_SIZE]>
implementations that provided access the the bytes in the page. Unfortunately,
this is not sound because PageRef doesn't have unique ownership over the page:
An aliasing page can be create by calling .clone() and the page is also aliased
when it's mapped into a page table. For these reasons, non-atomic accesses
could potentially cause a data race and those are UB. This leaves two
solutions for implementing the methods for PageRef:
1. Only use Relaxed atomic accesses. The codegen for this is generally quite
   bad because LLVM will access each byte individually.
2. Implement them in assembly. The upside of this is that if we control the
   assembly, so there's no compiler that will tell us what we can and can't do.
For this patch, I've opted for the second solution.

Signed-off-by: Tom Dohrmann <erbse.13@gmx.de>