kernel/insn_decode: Add emulation for Mov instruction variants
The emulation for Mov instruction variants is targeting for MMIO but not
memory. Introduced 3 new methods to the InsnMachineCtx. The method
translate_linear_addr is to facilitate transalting the virtual address
to the physical address. The method handle_mmio_read/write is to
facilitate MMIO accessing.

With these, the variants of moving general-purpose register/immediate to
memory is emulated by reading data from the register and writing to the
MMIO. The variants of moving memory to general-purpose register is
emulated by reading data from the MMIO and writing to the register.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>