kernel/syscall/obj: Add Obj trait and ObjHandle
The syscalls design philosophy is to provide a unified interface for the
user to access system resources, which are represented by object handles.
This makes objects a fundamental concept in the COCONUT-SVSM kernel

An object represents the type of resource like file, VM, vCPU in the
COCONUT-SVSM kernel which can be accessible by the user mode. Introduce
an Obj trait to represent such type of resource. The Obj trait can be
used to define the common functionalities of the objects. It is defined
with the trait bounds of Send and Sync, so the objects implementing Obj
trait could be sent to another thread and shared between threads safely.

ObjHandle is a unique identifier for an object in the current process.
An ObjHandle can be converted to a u32 id which can be used by the user
mode to access this object. The passed id from the user mode by the
syscalls can be converted to an `ObjHandle`, which can be used to access
the object in the COCONUT-SVSM kernel.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>