kernel/syscall/obj: Add object management
To facilitate the user mode using the object, the COCONUT-SVSM kernel
should:

- Manage the object's lifecycle properly. The underlying object should
  be dropped when the user mode closes the object handle via syscalls or
  the user mode is terminated without closing.

- Prevent one user mode process from misusing the object handle opened
  by another. But the object handles are shared among the threads within
  the same process.

To achieve the above goals, the opened object should be associated with
the process which creates it. The task structure is extended to hold the
created objects by a BTreeMap with the ObjHandle as the key and the
Arc<dyn Obj> as the value. It is wrapped by an Arc and protected by a
RWLock, to make it shared among the threads within the same process.

Three functions are introduced for the object management:
- obj_add() creates a new ObjHandle by adding the object to the current
  process.
- obj_close() removes and drops the object from the current process.
- obj_get() searches in the current process and returns the object if it
  exists.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>