syscall/obj: Define ObjHandle for the user mode
The object is exposed to the user mode via the object-opening related
syscalls, which returns the id of the object created by the COCONUT-SVSM
kernel. The user mode can make use this id to access the corresponding
object via other syscalls.

From the user mode's point of view, an ObjHanle is defined to wrap a u32
which is the value returned by an object-opening syscall. This u32 value
can be used as the input for the syscalls to access the corresponding
kernel object. The ObjHandle doesn't implement Copy/Clone trait, and the
From trait is implemented for a reference to ObjHandle to create a u32.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>