kernel/cpu: enable SMAP
Enable SMAP if supported.

If supported, SMAP is enabled if CR4.SMAP bit is set, and if RFLAGS.AC is
unset.
It means that we need to clear RFLAGS.AC in entries, to have the kernel
running with RFLAGS.AC = 0.

Two asm macros (asm_clac and asn_stac) have been created to respectively
clear and set RFLAGS.AC from assembly. Two Rust functions have also been
created to perform exactly the same but from Rust code. This is still unused
but it will will be useful when we'll have CPL3 support, to be able to
read/write in userspace (eg. in syscall handlers, or even with user #VC).

Signed-off-by: Thomas Leroy <thomas.leroy@suse.com>