cpu/idt: fix #HV handling path
When the #HV handler determines that it was entered from the
return-from-exception tail, it reverts to the previous stack frame and
proceeds to dispatch #HV events.  However, in this case, the code
neglected to reload the #HV doorbell page address, so the #HV event
handler would refer to random memory.

Signed-off-by: Jon Lange <jlange@microsoft.com>