kernel: fix test-in-svsm use of the test serial port
The test serial port must use the same platform abstraction as the rest
of the I/O port abstraction.  Since serial port construction is now
platform-specific, the construction is moved to the first use of the
test serial port.  The serial port cell is changed to an `Option` to
simplify the synchronization of the dynamic initialization across
consumers.

Also, for better abstraction, the operation to obtain the test serial
port doesn't emit a byte; it simply returns the port so the caller can
write whatever it wants.

Signed-off-by: Jon Lange <jlange@microsoft.com>