cpu: use Hyper-V construct structure to represent initial CPU context
Hyper-V defines a context structure that describes the initial context
to use when starting a virtual processor.  This structure is suitable as
a cross-platform representation of the initial context to use for CPU
startup which can be used in the platform abstraction layer to describe
the initial context for AP startup.

Signed-off-by: Jon Lange <jlange@microsoft.com>