kernel/idt: Ensure #HV handler RIP checks work when interrupts are enabled
The #HV handler currently checks EFLAGS.IF=1 at the start of the handler
and immediately processes the #HV events if enabled without performing
checks to see if RIP is within the VMPL switch or the iret window. The
RIP windows are only checked if EFLAGS.IF=0.

This means that if interrupts are enabled during either of these windows
and a #HV occurs then the behaviour is undefined.

The code has been reorganised in this patch to examine EFLAGS.IF only
after both RIP windows have been checked.

Signed-off-by: Roy Hopkins <roy.hopkins@suse.com>