requests: disable interrupts when reentering guest VMPL
When reentering the guest VMPL, no interrupts can be permitted once the
SVSM begins reevaluating state that is presented to the guest OS upon
reentry (such as guest APIC state), since handling an interrupt in this
window may alter and corrupt the guest state being prepared.

Note that it is no longer necessary to explicitly check for #HV doorbell
events in this flow because the use of #HV doorbell, which necessarily
implies the use of Restricted Injection, causes the SVSM to run
generally with interrupts enabled, and thus #HV doorbell events will
naturally be processed any time interrupts are enabled, or when they are
explicitly reenabled when the last `IrqGuard` is dropped.

Signed-off-by: Jon Lange <jlange@microsoft.com>