elf: Fix elf program header verification logic
In the current ELF parser implementation, `verify_phdr` function
returns an error when `p_vaddr` is not aligned. But since spec
doesn't guarantee that `p_vaddr` must be aligned, fixed the check
in our ELF parser. Also, a value of 0 or 1 for `p_align` means
no alignment is required, so fix this check as well.

Signed-off-by: Vijay Dhanraj <vijay.dhanraj@intel.com>
Signed-off-by: Joerg Roedel <jroedel@suse.de>