panic: use a platform-specific halt routine
Not all platforms support HLT as a way to halt execution, and therefore
the panic halt must be performed through a platform-specific routine.
However, at the time of a panic, the global platform object may not yet
be initialized.  Instead, use a more primitive mechanism for choosing a
platform-specific halt routine during panic.

Signed-off-by: Jon Lange <jlange@microsoft.com>