libmstpm: store generated bindings in OUT_DIR
Generated code shouldn't be stored directly in the src directory. All
files generated by build-scripts should be stored in the directory
pointed to in the OUT_DIR environment variable.
One advantage of this is that we can tell the build-script to emit
rerun-if statements that tell cargo to only rerun if any of the files
in this directory have changed. If we generate files into this
directory, cargo may not always realize that the generated files
shouldn't trigger a re-run.

Signed-off-by: Tom Dohrmann <erbse.13@gmx.de>