libtcgtpm: little hack to disable pthread check
The pthread library is only required for the simulator, but we
don't build it, so let's use this workaround to make the `configure`
happy and avoid checks that the pthread is available.

When the  https://github.com/TrustedComputingGroup/TPM/pull/7 will be
merged we can revert this commit and call `./configure --disable-pthread`.

With this patch we can then avoid using our fork in Coconut and switch
to using the upstream version directly (in the next commit).

Signed-off-by: Stefano Garzarella <sgarzare@redhat.com>