stage1: Rustify
Move stage1/*.S to a Rust crate. This allows for better interfacing with
other crates and reduces the amount of hard-coding and chances of
breakage.

Replace label-based relative addressing with raw pointer arithmetics to
avoid the "relocation truncated to fit" errors caused by the use of
relocation type R_X86_64_32S in 32-bit code.

Signed-off-by: Peter Fang <peter.fang@intel.com>