workspace: enable `unsafe_op_in_unsafe_fn` lint
It is good to mark functions as `unsafe` when the safety of their
behavior is not guaranteed.  It is not goot to permit these functions to
perform additional unsafe operations without warning.  All uses of
unsafe code should require explicit `unsafe` annotation regardless of
whether they appear in safe or unsafe functions.

Signed-off-by: Jon Lange <jlange@microsoft.com>