stage1: Increase stage2 area
Stage2 has gotten larger in size so increase the size of the stage2 area
to 800K (0x808000 - 0x8D0000).

Also, do sanity check in stage1 to make sure the stage2 image size
doesn't exceed 800K when loading stage2.

The memory regions are still hard-coded in gen_meta.c. This can be
tackled in the future if we decide to export these Rust constants.

Signed-off-by: Peter Fang <peter.fang@intel.com>