task/schedule: Split up create_user_task()
The create_user_task() function did the task creation and already put
the new task on the global TASKLIST and the cpu run-queue.

This is problematic because the task returned by create_user_task() is
not ready to run yet. When the ELF loading fails and exec_user()
returns an error, the task remains on the run-queue and TASKLIST. The
next call to schedule() will then try to run it, which can cause
undefined behavior in user-space.

So split the function up and only add a new user task to the run-queue
when it is fully initialized.

Signed-off-by: Joerg Roedel <jroedel@suse.de>