kernel/mm: Implement API for global mappings
Implement an API to establish mappings in the global part of the
COCONUT kernel address space. The mappings created with this API will
be accessible from all CPUs and guarded by a new struct, which takes
care of the unmapping once its life-time ends.

Place the global mappings in the unused range which mapped shared
stacks in the past.

Signed-off-by: Joerg Roedel <jroedel@suse.de>